style: 인재상 이미지 변경
