+++ src/Recruit/components/BodySection.tsx
@@ -1,11 +1,11 @@
 import styled from "styled-components";
 import Circle from "../assets/Circle.svg?react";
-import Arrow from "../../common/assets/arrow_down.svg?react";
-import { skillDataWithIcons, partDataWithIcons, stepData } from "../constants";
+import { skillDataWithIcons, stepData } from "../constants";
 import { useCountDownStore } from "../../common/components/CountDown";
 import { motion } from "framer-motion";
 import media from "../../common/styles/media";
 import RoadmapImg from "../assets/roadmap.png";
+import PartBox from "./PartBox";
 
 const Body = () => {
   const { isExpired } = useCountDownStore();
@@ -21,10 +21,9 @@ const Body = () => {
       <BoxContainer>
         <Title>숭실대 멋쟁이사자처럼은 이런 사람을 원해요</Title>
         <CardSection>
-          {skillDataWithIcons.map(({ id, icon: Icon, title, text }) => (
+          {skillDataWithIcons.map(({ id, icon, title, text }) => (
             <SkillCard key={id}>
-              {Icon ? <SkillIcon as={Icon} /> : null}
-              {/*아이콘 있을 때만 렌더링 되도록 */}
+              <SkillIcon src={icon} alt={title} />
               <Text>
                 <TextBody4>{title}</TextBody4>
                 <CardText>{text}</CardText>
@@ -34,21 +33,7 @@ const Body = () => {
         </CardSection>
       </BoxContainer>
 
-      {/* 모집 파트 */}
-      <BoxContainer>
-        <Title>모집 파트</Title>
-        <PartSection>
-          {partDataWithIcons.map(({ id, name, icon: Icon }) => (
-            <PartCard key={id}>
-              <ArrowIC />
-              <PartBox>
-                <PartName>{name}</PartName>
-                {Icon ? <PartIcon as={Icon} /> : null}
-              </PartBox>
-            </PartCard>
-          ))}
-        </PartSection>
-      </BoxContainer>
+      <PartBox />
 
       {/* 로드맵 */}
       <BoxContainer>
@@ -147,6 +132,7 @@ const SkillCard = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
+  padding-top: 5rem;
 
   width: 32rem;
   height: 40rem;
@@ -166,10 +152,11 @@ const SkillCard = styled.div`
     border-radius: 5.94px;
     display: flex;
     flex-direction: row;
+    padding-top:0;
   `};
 `;
 
-const SkillIcon = styled.div`
+const SkillIcon = styled.img`
   width: 20rem;
   height: 20rem;
 
@@ -206,82 +193,6 @@ const CardText = styled.p`
   ${({ theme }) => media.small`
     ${theme.mixins.font(theme.fonts.Pretendard.body7)}
   `};
-`;
-
-// 파트
-const PartSection = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 3rem;
-  flex-wrap: wrap;
-
-  ${media.small`
-    flex-wrap: wrap;
-    gap: 1rem;
-  `};
-`;
-
-const PartCard = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: end;
-  justify-content: center;
-
-  width: 30rem;
-  height: 20rem;
-  flex-shrink: 0;
-  border-radius: 10px;
-  padding: 2rem;
-
-  background: linear-gradient(
-    180deg,
-    rgba(255, 255, 255, 0.1) 0%,
-    rgba(255, 255, 255, 0.08) 55.49%,
-    rgba(255, 255, 255, 0.04) 100%
-  );
-
-  ${media.small`
-    width: 16rem;
-    height: 11rem;
-  `};
-`;
-
-const PartName = styled.p`
-  text-align: left;
-  font-feature-settings: "calt" off;
-
-  ${({ theme }) => theme.mixins.font(theme.fonts.Suit.body1)}
-  line-height: normal;
-
-  ${({ theme }) => media.small`
-    ${theme.mixins.font(theme.fonts.Suit.body5)}
-  `};
-`;
-
-const PartIcon = styled.div`
-  width: 14.5rem;
-  height: 14.5rem;
-
-  ${media.small`
-    width: 6rem;
-    height: 6rem;
-  `};
-`;
-
-const ArrowIC = styled(Arrow)`
-  width: 1.6rem;
-  height: 0.9rem;
-
-  ${media.small`
-    width: 2rem;
-    height: 2rem;
-  `};
-`;
-
-const PartBox = styled.div`
-  display: flex;
-  align-items: end;
 `;
 
 // 로드맵
